refactor(subagent-tree): extract modules, normalize line breaks
- Extract TerminalWriter seam into new terminal-writer.ts module
- Extract runtime-detection helpers into new runtime-detection.ts module with re-exports
- Normalize all line breaks to \r\n in PseudoterminalTerminalWriter.write() to fix multi-line rendering staircase
- Reduce command-runtime.ts from 570 to 368 lines, under the 500-line limit
- Add coverage thresholds and multi-line unit tests for new modules

Refs: #325
