+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/build.2026-07-06T23-56.md
@@ -1,0 +1,10 @@
+# Phase 2 — Build (P2-T5)
+
+**Timestamp:** 2026-07-06T23-56
+**Command:** `npm run build` (from `extensions/drm-copilot/`)
+**EXIT_CODE:** 0
+**Output Summary:** `tsc -p ./ --noEmit` succeeded with no output (0 errors),
+followed by `bundle:extension` (`esbuild-extension.cjs`) and
+`bundle:mcp-server` (`esbuild-mcp-server.cjs`), both completing without
+errors. The new `src/terminal-writer.ts` module and its import from
+`src/subagent-tree-command.ts` bundle cleanly.
