+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/build.2026-07-07T04-02.md
@@ -1,0 +1,8 @@
+Timestamp: 2026-07-07T04-02
+Command: npm run build
+EXIT_CODE: 0
+Output Summary: `tsc -p ./ --noEmit` succeeded, followed by `bundle:extension`
+(`node esbuild-extension.cjs`) and `bundle:mcp-server` (`node esbuild-mcp-server.cjs`),
+both completing without error. The extension bundle successfully resolves the new
+`src/runtime-detection.ts` module and the updated re-exports in
+`src/command-runtime.ts`.
