+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/file-size-verification-final.2026-07-06T23-58.md
@@ -1,0 +1,26 @@
+# Phase 2 — Final File-Size Verification (P2-T6)
+
+**Timestamp:** 2026-07-06T23-58
+**Command:** `wc -l extensions/drm-copilot/src/command-runtime.ts extensions/drm-copilot/src/terminal-writer.ts`
+**EXIT_CODE:** 0
+**Output Summary:**
+```
+570 extensions/drm-copilot/src/command-runtime.ts
+100 extensions/drm-copilot/src/terminal-writer.ts
+```
+
+`terminal-writer.ts` (100 lines) satisfies the `<= 500` acceptance criterion.
+
+`command-runtime.ts` (570 lines) does **not** satisfy the plan's `<= 500`
+acceptance criterion, unchanged from the P1-T3 measurement taken immediately
+after the extraction (the Phase-2 toolchain loop — format/lint/typecheck/
+test:coverage/build — makes no further edits to `command-runtime.ts`). See
+`evidence/qa-gates/file-size-verification.2026-07-06T23-44.md` for the
+root-cause analysis: `command-runtime.ts` was already 531 lines on `main`
+before this feature (pre-existing, out-of-scope per the remediation's "Do
+Not Do" list), and this remediation's sole authorized extraction (the
+`TerminalWriter` seam) returns the file to 531 + 39 (`getClaudeProjectsRoot`,
+retained per plan task P1-T4) = 570 lines, not to `<= 500`. This is recorded
+as an unresolved discrepancy between the plan's literal acceptance
+criterion and the scope authorized by the "Do Not Do" list; see the
+executor's final completion report for the escalation.
