+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/file-size-verification-final.2026-07-07T04-02.md
@@ -1,0 +1,14 @@
+Timestamp: 2026-07-07T04-02
+Command: wc -l extensions/drm-copilot/src/command-runtime.ts && wc -l extensions/drm-copilot/src/runtime-detection.ts && wc -l extensions/drm-copilot/src/terminal-writer.ts
+EXIT_CODE: 0
+Output Summary: Final line counts after the clean five-stage toolchain pass
+(format -> lint -> typecheck -> test:coverage -> build): `command-runtime.ts` = 368
+lines, `runtime-detection.ts` = 211 lines, `terminal-writer.ts` = 100 lines. All three
+are <= 500. `command-runtime.ts` satisfies Fix 1's stated goal ("returns to at or
+under 500 lines") via the additional authorized extraction of the
+executable/runtime-resolution group (private helpers `executableExists`,
+`getPathExtensions`, `findExecutableOnPath`, `normalizeExecutablePath`,
+`buildCodexExecutableCandidates`, `findCodexInInstalledExtensionRoots`, plus exported
+`resolveCodexExecutable`/`detectRuntime` and the `RuntimeKind`/`RuntimeResolution`
+types) into `src/runtime-detection.ts`, with re-exports preserved in
+`command-runtime.ts` for `src/extension.ts` and `test/extension-test-harness.ts`.
