+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/file-size-verification.2026-07-07T04-02.md
@@ -1,0 +1,12 @@
+Timestamp: 2026-07-07T04-02
+Command: wc -l extensions/drm-copilot/src/command-runtime.ts && wc -l extensions/drm-copilot/src/runtime-detection.ts && wc -l extensions/drm-copilot/src/terminal-writer.ts
+EXIT_CODE: 0
+Output Summary: Post-extraction line counts (additional authorized extraction of the
+executable/runtime-resolution group into `src/runtime-detection.ts`, completing Fix 1
+for issue #325): `command-runtime.ts` = 368 lines, `runtime-detection.ts` = 211 lines,
+`terminal-writer.ts` = 100 lines. `command-runtime.ts` is <= 500 (target met; was 570
+lines before this extraction, 531 lines pre-existing on main). `runtime-detection.ts`
+is <= 500. `terminal-writer.ts` is <= 500 (unchanged from prior remediation cycle).
+This satisfies the P1-T3 acceptance criterion (`command-runtime.ts` <= 500 and
+`terminal-writer.ts` <= 500) and additionally verifies the new
+`runtime-detection.ts` module.
