+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/final-qc-clean-pass.2026-07-06T23-59.md
@@ -1,0 +1,38 @@
+# Phase 2 — Final QC Clean-Pass Summary (P2-T7)
+
+**Timestamp:** 2026-07-06T23-59
+
+P2-T1 through P2-T6 all executed in a single pass with no restart. Stage
+exit codes and referenced evidence artifacts:
+
+| Task | Command | EXIT_CODE | Evidence Artifact |
+|------|---------|-----------|--------------------|
+| P2-T1 | `npm run format` | 0 | `format.2026-07-06T23-52.md` |
+| P2-T2 | `npm run lint` | 0 | `lint.2026-07-06T23-52.md` |
+| P2-T3 | `npm run typecheck` | 0 | `typecheck.2026-07-06T23-52.md` |
+| P2-T4 | `npm run test:coverage` | 0 | `test-coverage.2026-07-06T23-55.md` |
+| P2-T5 | `npm run build` | 0 | `build.2026-07-06T23-56.md` |
+| P2-T6 | `wc -l ...command-runtime.ts ...terminal-writer.ts` | 0 | `file-size-verification-final.2026-07-06T23-58.md` |
+
+All six commands above completed with `EXIT_CODE: 0`, and none of P2-T1
+through P2-T5 modified any file (a re-run of `npm run format` after the
+pass confirmed every scanned file remained "(unchanged)"), so the five-stage
+toolchain loop (`general-code-change.md`'s "Mandatory Toolchain Loop")
+completed in a single clean pass with no restart required.
+
+No acceptance-criteria checkbox in `issue.md` was modified by this
+remediation. `issue.md` was not edited by any task in this plan; all ten
+criteria remain in the state recorded before this remediation cycle began
+(all ten already checked `[x]`, per `remediation-inputs.2026-07-07T03-30.md`'s
+"Trigger reason": "All ten acceptance criteria in `issue.md` pass on their
+literal wording").
+
+**Outstanding discrepancy (not a toolchain-stage failure):** P2-T6's own
+numeric acceptance criterion ("both counts are <= 500") is not met for
+`extensions/drm-copilot/src/command-runtime.ts` (570 lines). This is a
+pre-existing condition documented in
+`evidence/qa-gates/file-size-verification.2026-07-06T23-44.md` and
+`evidence/qa-gates/file-size-verification-final.2026-07-06T23-58.md`, not a
+failure of any of the five toolchain commands above (all five exited 0 and
+changed no files). See the executor's final completion report for the
+escalation of this discrepancy.
