+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/final-qc-clean-pass.2026-07-07T04-02.md
@@ -1,0 +1,25 @@
+Timestamp: 2026-07-07T04-02
+Command: (summary artifact; no single command)
+EXIT_CODE: 0
+Output Summary: This remediation sub-cycle completed the authorized additional
+extraction required for Fix 1 (issue #325) — moving the executable/runtime-resolution
+group out of `src/command-runtime.ts` into a new `src/runtime-detection.ts` module —
+and re-ran the full five-stage toolchain from `extensions/drm-copilot/`. Format
+(P2-T1 equivalent, this cycle's `format.2026-07-07T04-02.md`) auto-fixed one file
+(`src/command-runtime.ts`, an import-wrap); the loop was restarted and re-verified
+clean via `npx prettier --check`. Lint, typecheck, test:coverage, and build then all
+passed with `EXIT_CODE: 0` on the first subsequent attempt, with no further file
+changes, completing the loop in a single clean pass after the one format auto-fix
+restart:
+
+- format.2026-07-07T04-02.md — EXIT_CODE: 0 (1 file auto-fixed, then re-verified clean)
+- lint.2026-07-07T04-02.md — EXIT_CODE: 0
+- typecheck.2026-07-07T04-02.md — EXIT_CODE: 0
+- test-coverage.2026-07-07T04-02.md — EXIT_CODE: 0 (1531/1531 tests passed; all five
+  named files >= 85% lines / >= 75% branches)
+- build.2026-07-07T04-02.md — EXIT_CODE: 0
+- file-size-verification-final.2026-07-07T04-02.md — EXIT_CODE: 0
+  (`command-runtime.ts` = 368 lines, `runtime-detection.ts` = 211 lines,
+  `terminal-writer.ts` = 100 lines; all <= 500)
+
+No acceptance-criteria checkbox in `issue.md` was modified by this remediation.
