+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/format.2026-07-06T23-52.md
@@ -1,0 +1,10 @@
+# Phase 2 — Format (P2-T1)
+
+**Timestamp:** 2026-07-06T23-52
+**Command:** `npm run format` (from `extensions/drm-copilot/`)
+**EXIT_CODE:** 0
+**Output Summary:** Prettier reported every scanned file (all `src/**/*.ts`,
+`test/**/*.ts`, `*.json`, `*.cjs`, including the newly added
+`src/terminal-writer.ts` and `test/terminal-writer.test.ts`) as
+"(unchanged)". No files were rewritten by the formatter; loop proceeds to
+lint without restart.
