+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/format.2026-07-07T04-02.md
@@ -1,0 +1,9 @@
+Timestamp: 2026-07-07T04-02
+Command: npm run format
+EXIT_CODE: 0
+Output Summary: Prettier reformatted 1 file (`src/command-runtime.ts` — wrapped the
+multi-symbol `import { detectRuntime, type RuntimeKind, type RuntimeResolution } from
+"./runtime-detection";` line onto multiple lines). Per the toolchain rerun rule, the
+loop was restarted from format after this auto-fix; a follow-up
+`npx prettier --check src/runtime-detection.ts src/command-runtime.ts jest.config.cjs`
+confirmed "All matched files use Prettier code style!" with no further changes.
