+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/lint.2026-07-06T23-52.md
@@ -1,0 +1,10 @@
+# Phase 2 — Lint (P2-T2)
+
+**Timestamp:** 2026-07-06T23-52
+**Command:** `npm run lint` (from `extensions/drm-copilot/`)
+**EXIT_CODE:** 0
+**Output Summary:** `eslint --no-error-on-unmatched-pattern src test` produced
+no output (0 errors, 0 warnings) across the full `src`/`test` tree, including
+the new `src/terminal-writer.ts`, `test/terminal-writer.test.ts`, and the
+edited `src/command-runtime.ts`, `src/subagent-tree-command.ts`,
+`test/command-runtime.test.ts`, `test/subagent-tree-command.test.ts`.
