+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/lint.2026-07-07T04-02.md
@@ -1,0 +1,6 @@
+Timestamp: 2026-07-07T04-02
+Command: npm run lint
+EXIT_CODE: 0
+Output Summary: `eslint --no-error-on-unmatched-pattern src test` completed with no
+reported errors or warnings across `src` and `test`, including the new
+`src/runtime-detection.ts` module and the edited `src/command-runtime.ts`.
