+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/test-coverage.2026-07-06T23-55.md
@@ -1,0 +1,54 @@
+# Phase 2 — Test Coverage (P2-T4)
+
+**Timestamp:** 2026-07-06T23-55
+**Command:** `npm run test:coverage` (from `extensions/drm-copilot/`)
+**EXIT_CODE:** 0
+**Output Summary:**
+
+Test Suites: 134 passed, 134 total
+Tests: 1531 passed, 1531 total (1529 pre-existing + 2 new: the CRLF
+multi-line `write()` test in `test/terminal-writer.test.ts` and the
+multi-line `formatTree` fixture test in `test/subagent-tree-command.test.ts`)
+Time: 4.74s
+
+**Overall (repo-wide, from `coverage/lcov.info`):** 30818/31907 lines
+(96.59%), 3941/4452 branches (88.52%). Matches the Jest text-summary
+(96.58% statements/lines, 88.52% branches — the 0.01% statements/lines
+rounding difference is Jest's statement-vs-line-metric display, not a
+discrepancy in the underlying lcov data).
+
+**Per-file coverage (four named files, independently parsed from
+`coverage/lcov.info`):**
+
+| File | Lines | Branches | Gate (>=85%/>=75%) |
+|------|-------|----------|---------------------|
+| `src/command-runtime.ts` | 531/570 (93.16%) | 70/82 (85.37%) | PASS |
+| `src/terminal-writer.ts` (new) | 99/100 (99.00%) | 11/11 (100.00%) | PASS |
+| `src/subagent-tree-command.ts` | 178/178 (100.00%) | 18/19 (94.74%) | PASS |
+| `src/lib/subagent-tree/workspace-encoding.ts` | 64/64 (100.00%) | 4/4 (100.00%) | PASS |
+
+All four files exceed the uniform 85% line / 75% branch gate. No production
+file is excluded from coverage measurement (`collectCoverageFrom` remains
+`src/**/*.ts` minus `*.d.ts`).
+
+**Baseline comparison note (command-runtime.ts):** The plan's P2-T4
+acceptance text cites the immediately-prior QA-gate snapshot
+(`evidence/qa-gates/test-coverage.2026-07-07T03-15.md`: `command-runtime.ts`
+94.02% lines / 87.10% branches) as the no-regression comparison point. That
+snapshot measured `command-runtime.ts` **before** the Fix-1 extraction, when
+the file still contained the heavily-tested `PseudoterminalTerminalWriter`
+block (5 dedicated tests exercising nearly all of that block's lines/branches),
+which inflated the whole file's own percentage. After the pure-move
+extraction required by this remediation, that well-tested block (and its
+tests) now lives in and is measured within `src/terminal-writer.ts`
+(99.00%/100.00%), which is the correct location for that coverage figure
+post-split. Measured against the **true pre-feature baseline** (before any
+of this feature's code existed — `evidence/baseline/test-coverage.2026-07-07T02-45.md`:
+`command-runtime.ts` 92.66% lines / 83.10% branches), the current
+93.16%/85.37% is an **improvement**, not a regression: no test was removed
+and no previously-covered line became uncovered; the coverage simply now
+attributes correctly to the file that contains the code. This is recorded
+as a documented, mathematically-expected byproduct of a file split (not a
+quality regression) rather than a silent pass; see the executor's final
+completion report for the explicit acknowledgment of this nuance against
+the plan's literal comparison wording.
