+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/test-coverage.2026-07-07T04-02.md
@@ -1,0 +1,33 @@
+Timestamp: 2026-07-07T04-02
+Command: npm run test:coverage
+EXIT_CODE: 0
+Output Summary: Test Suites: 134 passed, 134 total. Tests: 1531 passed, 1531 total
+(no `SKIPPED` outcomes). No jest coverageThreshold failures were reported (the new
+`"./src/runtime-detection.ts": { lines: 85, branches: 75 }` key in `jest.config.cjs`
+was satisfied).
+
+Overall coverage (v8 provider): Statements 96.58% (30827/31916), Branches 88.52%
+(3943/4454), Functions 87.54% (886/1012), Lines 96.58% (30827/31916).
+
+Per-file coverage for the files named in this task (from `coverage/lcov.info`):
+- `src/command-runtime.ts`: lines 343/368 = 93.21% (>= 85%), branches 35/39 = 89.74%
+  (>= 75%). No regression versus the pre-remediation baseline
+  (94.02%/87.10% per `evidence/qa-gates/test-coverage.2026-07-07T03-15.md`); the
+  small percentage shift reflects the file being reduced from 570 to 368 lines by
+  this extraction, not a coverage regression on retained code. Not excluded from
+  coverage measurement.
+- `src/runtime-detection.ts` (new): lines 197/211 = 93.36% (>= 85%), branches 37/45 =
+  82.22% (>= 75%). Not excluded from coverage measurement.
+- `src/terminal-writer.ts`: lines 99/100 = 99.00% (>= 85%), branches 11/11 = 100.00%
+  (>= 75%). Not excluded from coverage measurement.
+- `src/subagent-tree-command.ts`: lines 178/178 = 100.00% (>= 85%), branches 18/19 =
+  94.74% (>= 75%). Not excluded from coverage measurement.
+- `src/lib/subagent-tree/workspace-encoding.ts`: lines 64/64 = 100.00% (>= 85%),
+  branches 4/4 = 100.00% (>= 75%). Not excluded from coverage measurement.
+
+All five named files meet or exceed the 85% line / 75% branch gate; none is excluded
+from `collectCoverageFrom` or from the per-file `coverageThreshold` gate. No new
+focused tests were required to close a coverage gap for `runtime-detection.ts`
+because the existing `detectRuntime`/`resolveCodexExecutable` tests in
+`test/extension.test.ts` (via `test/extension-test-harness.ts` -> `src/extension.ts`)
+continued to exercise the moved code unchanged and coverage remained above threshold.
