+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/typecheck.2026-07-06T23-52.md
@@ -1,0 +1,9 @@
+# Phase 2 — Typecheck (P2-T3)
+
+**Timestamp:** 2026-07-06T23-52
+**Command:** `npm run typecheck` (from `extensions/drm-copilot/`)
+**EXIT_CODE:** 0
+**Output Summary:** `tsc -p ./ --noEmit` produced no output (0 type errors).
+The new `TerminalWriter` import in `subagent-tree-command.ts` and
+`subagent-tree-command.test.ts`/`terminal-writer.test.ts` type-check cleanly
+against `src/terminal-writer.ts`.
