+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/qa-gates/typecheck.2026-07-07T04-02.md
@@ -1,0 +1,8 @@
+Timestamp: 2026-07-07T04-02
+Command: npm run typecheck
+EXIT_CODE: 0
+Output Summary: `tsc -p ./ --noEmit` completed with 0 type errors, confirming the
+one-directional import of `RuntimeKind`/`RuntimeResolution`/`detectRuntime` from the
+new `src/runtime-detection.ts` module into `src/command-runtime.ts`, and the
+re-exports from `command-runtime.ts`, type-check cleanly for all consumers
+(`src/extension.ts`, `test/extension-test-harness.ts`, `test/extension.test.ts`).
