+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/remediation-baseline/audit-inputs-read.md
@@ -1,0 +1,23 @@
+# Phase 0 — Audit Inputs Read (P0-T2)
+
+**Timestamp:** 2026-07-06T23-44
+
+**Files read and confirmation:**
+
+1. `docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/policy-audit.2026-07-07T03-30.md`
+   — Read in full. Confirms two Blocking findings: `command-runtime.ts` at
+   669 lines (exceeds the 500-line limit) and `PseudoterminalTerminalWriter.write()`
+   not normalizing internal `\n` line breaks to `\r\n`.
+2. `docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/code-review.2026-07-07T03-30.md`
+   — Read (Blocking findings section confirmed). Recommends "Needs Revision"
+   pending the same two fixes: CRLF normalization and file-size extraction.
+3. `docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/feature-audit.2026-07-07T03-30.md`
+   — Read (criterion #5 notes confirmed). All ten acceptance criteria pass on
+   literal wording; the two Blocking findings are code-quality/policy issues
+   outside the literal AC text.
+4. `docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/remediation-inputs.2026-07-07T03-30.md`
+   — Read in full. Enumerates exactly the two fixes (TerminalWriter
+   extraction; CRLF normalization) and the binding "Do Not Do" list.
+
+All four file paths above were read before editing began in this
+remediation cycle.
