+++ docs/features/active/2026-07-06-fix-subagent-tree-discovery-terminal-325/evidence/remediation-baseline/line-count-baseline.md
@@ -1,0 +1,9 @@
+# Phase 0 — Pre-Fix Line Count Baseline (P0-T3)
+
+**Timestamp:** 2026-07-06T23-44
+**Command:** `wc -l extensions/drm-copilot/src/command-runtime.ts`
+**EXIT_CODE:** 0
+**Output Summary:** `669 extensions/drm-copilot/src/command-runtime.ts` — matches the
+expected pre-fix length recorded in `remediation-plan.2026-07-07T03-30.md`
+(exactly 669), confirming the file-size Blocking finding prior to the
+Fix 1 extraction.
